requests할 때 오류났을 때 재귀함수로 처리
- requests할 때 time.sleep()을 해줘도 오류나는 경우가 많다.
- 오류났을 때 재귀함수로 처리해서 될 때까지 함수를 반복한다
