+++ crawling/src/dcinside_content_crawler.py
@@ -3,7 +3,7 @@
 # 20230927
 #############################
 
-from crawling_tool import get_gall_id, get_driver
+from crawling_tool import get_gall_id, get_new_row_from_main_content, get_reply_list, is_ignore_reply, get_reply_date
 import utility_module as util
 import requests
 import pandas as pd
@@ -15,59 +15,6 @@ from bs4 import BeautifulSoup
 header = {
         "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36"
     }
-
-
-#####################################
-# 본문에서 new_row를 얻어오는 함수
-def get_new_row_from_main_content(time_sleep_sec, url_row):
-    is_comment = 0  # 본문이므로 0
-    response = requests.get(url_row['url'], headers=header)
-    time.sleep(time_sleep_sec)
-    soup = BeautifulSoup(response.text, "html.parser")
-    content = util.preprocess_content_dc(soup.find('div', {"class": "write_div"}).text)
-    content = url_row['title'] + " " + content
-    new_row = [url_row['date'], url_row['title'], url_row['url'], url_row['media'], content, is_comment]
-    return new_row
-
-
-#####################################
-# 기능 : url을 받아 reply_list를 리턴합니다
-# 리턴값 : reply_list
-def get_reply_list(time_sleep_sec, url):
-    driver = get_driver()
-    driver.get(url)
-    time.sleep(time_sleep_sec)
-    soup = BeautifulSoup(driver.page_source, "html.parser")
-    reply_list = soup.find_all("li", {"class": "ub-content"})
-    return reply_list
-
-
-#####################################
-# 기능 : 댓글 html코드를 받아서, 댓글의 date를 리턴합니다
-# 리턴값 : 2023-10-06 형식의 문자열
-def get_reply_date(reply):
-    temp_date = reply.find("span", {"class": "date_time"}).text.replace(".", "-")  # 댓글 등록 날짜 추출
-    if temp_date[:2] == "20":  # 작년 이전은 "2022.09.07 10:10:54" 형식임
-        date = temp_date[:10]
-    else:  # 올해는 "09.30 10:10:54" 형식임
-        date = str(datetime.datetime.now().year) + "-" + temp_date[:5]  # 올해 년도를 추가함
-    return date
-
-
-#####################################
-# 기능 : 무시해야하는 댓글이면, True를 반환하고, 필요한 댓글이면 False를 반환합니다
-def is_ignore_reply(reply):
-    if reply.select_one("p.del_reply"):
-        print("[삭제된 코멘트입니다]")
-        return True
-    elif reply.find('span', {'data-nick': '댓글돌이'}):
-        print("[댓글돌이는 무시합니다]")
-        return True
-    elif reply.find('div', {'class': 'comment_dccon'}):
-        print("[디시콘은 무시합니다]")
-        return True
-    else:
-        return False
 
 
 #####################################
@@ -109,23 +56,10 @@ def get_content_dc(gall_url):
         print(f"[index : {index}] {url_row['url']}")
 
         # {step 1} 본문 정보 row를 data_list에 추가
-        try:
-            print("{step 1} 본문 정보를 추가하겠습니다")
-            print("1초 기다리기")
-            new_row = get_new_row_from_main_content(1, url_row)  # new_row에 정보를 채워둔다
-            data_list.append(new_row)   # data_list에 new_row를 추가한다
-            print(f"[본문을 추가했습니다] {new_row}")
-        except Exception as e:
-            # 오류뜨는 것들에 한해, 시간을 넉넉하게 줘서 오류를 방지하고 데이터의 무결성을 유지한다
-            print("1초가 너무 짧아서, 10초 기다리겠습니다")
-            try:
-                new_row = get_new_row_from_main_content(10, url_row)  # new_row에 정보를 채워둔다
-                data_list.append(new_row)  # data_list에 new_row를 추가한다
-                print(f"[본문을 추가했습니다] {new_row}")
-            except Exception as e:
-                status = "[{step 1} 본문 정보 row를 data_list에 추가]"
-                print(f'[ERROR][index : {index}]{status}[error message : {e}]')
-                error_log.append([index, status, e, url_row['title'], url_row['url']])
+        print("{step 1} 본문 정보를 추가하겠습니다")
+        new_row = get_new_row_from_main_content(1, url_row)  # new_row에 정보를 채워둔다
+        data_list.append(new_row)                                         # data_list에 new_row를 추가한다
+        print(f"[본문을 추가했습니다] {new_row}")
 
         # {step 2} 댓글들 정보들을 불러오겠습니다
         # new_row 형식 : ['date', 'title', 'url', 'media', 'content', 'is_comment']
