+++ crawling/src/dcinside_url_crawler.py
@@ -23,6 +23,7 @@ header = {
 def get_url_dcinside(gall_url, keyword):
     # 0. 기본값 세팅 단계
     try:
+        keyword = util.convert_to_unicode(keyword)          # 입력받은 키워드를 유니코드로 변환한다
         gall_id = get_gall_id(gall_url)                     # 갤러리 id
         url_base = get_url_base(gall_url)                   # "https" 부터 "board/" 이전까지의 url 부분 (major갤, minor갤, mini갤)
         max_num = get_max_num(keyword, gall_id, url_base)   # 검색결과 중, 가장 큰 글번호 10000단위로 올림한 값/10000
@@ -44,16 +45,10 @@ def get_url_dcinside(gall_url, keyword):
 
     # 1. url 크롤링
     for search_pos in range(max_num, 0, -10000):
-        try:
-            print(f"[1만 개 단위로 검색합니다. 검색어 : {keyword} ] " + "*"*100)
-            print(f"[search_pos : {search_pos}] " + "*"*114)
-            temp_url = f"{url_base}/board/lists/?id={gall_id}&page=1&search_pos=-{search_pos}&s_type=search_subject_memo&s_keyword={keyword}"
-            last_page = get_last_page(temp_url)     # 1만 단위 검색결과의 마지막 페이지
-        except Exception as e:
-            step = "[url 크롤링]"
-            print(f'[error]{step}[error message : {e}]')
-            error_log.append([temp_url, step, e])
-            continue
+        print(f"[1만 개 단위로 검색합니다. 검색어 : {keyword} ] " + "*"*100)
+        print(f"[search_pos : {search_pos}] " + "*"*100)
+        temp_url = f"{url_base}/board/lists/?id={gall_id}&page=1&search_pos=-{search_pos}&s_type=search_subject_memo&s_keyword={keyword}"
+        last_page = get_last_page(temp_url)     # 1만 단위 검색결과의 마지막 페이지
         # 해외주식갤러리처럼 컨텐츠 많을때, 1,2,3,4,5 이렇게 페이징 되어있는거 있음. 이걸 페이지 넘기면서 크롤링
         for page in range(1, last_page+1):
             try:
@@ -65,9 +60,9 @@ def get_url_dcinside(gall_url, keyword):
                 element_list = soup.select("table.gall_list tr.ub-content")     # 한 페이지 전체 글 리스트
                 print(f"[글 개수 : {len(element_list)}]")
             except Exception as e:
-                step = "[검색결과 페이지 불러오기]"
-                print(f'[error]{step}[error message : {e}]')
-                error_log.append([search_url, step, e])
+                status = "[검색결과 페이지 불러오기]"
+                print(f'[error]{status}[error message : {e}]')
+                error_log.append([search_url, status, e])
                 continue
             # 글 하나씩 뽑아서 크롤링
             for element in element_list:    # element == 글 하나
@@ -82,9 +77,9 @@ def get_url_dcinside(gall_url, keyword):
                     data_list.append(new_row)  # data_list에 크롤링한 정보 저장
                     print(f"[page : {page}/{last_page}] new_row : {new_row}")
                 except Exception as e:
-                    step = "[검색결과에서 글 하나씩 크롤링]"
-                    print(f'[error]{step}[error message : {e}]')
-                    error_log.append([url, step, e])
+                    status = "[검색결과에서 글 하나씩 크롤링]"
+                    print(f'[error]{status}[error message : {e}]')
+                    error_log.append([url, status, e])
                     continue
 
     # 2. 파일로 저장
@@ -92,9 +87,9 @@ def get_url_dcinside(gall_url, keyword):
         df_result = pd.DataFrame(data_list, columns=['date', 'title', 'url', 'media'])
         util.save_file(df_result, folder_path, f"{file_name}.csv")
     except Exception as e:
-        step = "[파일 저장]"
+        status = "[파일 저장]"
         print('[error] ', e)
-        error_log.append([url, step, e])
+        error_log.append([url, status, e])
 
     # 3. 에러로그확인
     try:
