+++ crawling/tests/test_dcinside.py
@@ -1,7 +1,8 @@
 from dcinside_content_crawler import get_content_dc
+from dcinside_url_crawler import get_url_dcinside
 
 # 설정값
-keyword = "토스"
+keyword = "에스엠"
 blacklist = ["토스트", "도리토스", "치토스", "멘토스", "셀토스", "키보토스", "프로토스", "테스토스테론"]
 gall_url = {
     "해외주식": "https://gall.dcinside.com/mgallery/board/lists/?id=tenbagger",
@@ -17,12 +18,12 @@ gall_url = {
 
 # 해외주식, 코스피, 체크카드, SFF, 토스
 # [1. url 크롤링]
-# get_url_dcinside(gall_url["토스"], keyword)
+# get_url_dcinside(gall_url["코스피"], keyword)
 
 # get_url_dc_toss(keyword, gall_url["토스"])
 
 # [2. content 크롤링]
-get_content_dc(gall_url["신용카드"])
+get_content_dc(gall_url["코스피"])
 
 #########################################
 # url = "https://gall.dcinside.com/board/lists/?id=skwyverns_new1&s_type=search_subject_memo&s_keyword=.E3.85.87.E3.85.87"
